raspberrypi/requirements.txt
pyserial==3.5
simpleaudio==1.0.4
